Create livingroom lights based on lux.dz
Lights will be switched on in the livingroom when lux is below set value and based on the current mode. Light above dresser will turn on earlier then the other lights. When mode is set to 'Alles uit' (at night) and there is movement two lights will turn on for 2.5 minutes at 15%.